build(a5/fdwic): C.2 — std::atomic → volatile + __atomic_* on dist_engine
Replaces every std::atomic field / call with volatile T + __atomic_* GCC
intrinsics so the same source compiles under CCEC (no <atomic>, no
std::memory_order). A thin atom_load/store/cas_weak/cas_strong/fetch_add/
fetch_sub/thread_fence wrapper layer at the top of dist_engine.cpp keeps
the acq/rel/relaxed semantics readable at call sites; both host libstdc++
and the CCEC AICore backend lower these to the same fences trb already
uses (see trb aicore_executor / aicpu_executor for existing __atomic_*).

Fields participating in fetch_add promoted to int64_t (CCEC rejects s32
AtomicLoadAdd on GM addrspace 1, verified by the earlier aicore_executor
gate): DistGlobal::started_count, DistGlobal::replay_done, WonSlot::remaining,
Runtime::DistHandoff::done_count. Other fields keep their original width
(load/store/CAS work fine at 8/32/64 bits).

Sim regression green: benchmark_bgemm + paged_attention_unroll Case1 golden
PASS. CCEC still fails on the next layer (g_dist/g_self device-visibility
+ pto_types.h Arg::add_scalar_one / TaskOutputTensors ctor host-tagged
methods) — addressed by C.3 and subsequent stages.

+++ src/a5/runtime/fully_distributed_within_core/runtime/runtime.h
@@ -271,7 +271,10 @@ public:
         volatile uint64_t core_main_fn;  // DistCoreMainFn (in AICPU .so)
         volatile uint32_t go;            // 1 once engine wired and cores may start
         volatile int32_t num_workers;    // number of AICore workers participating
-        volatile int32_t done_count;     // workers atomically increment when done
+        // int64_t (not int32_t): CCEC's aicore backend refuses to lower
+        // AtomicLoadAdd on a 32-bit GM (addrspace 1) target, and workers do
+        // __atomic_add_fetch on this counter from aicore_executor.cpp.
+        volatile int64_t done_count;     // workers atomically increment when done
     } dist;
 
 private:
